+++ index.ts
@@ -574,6 +574,14 @@ namespace Enumerable {
         indexOf<U>(item: U,
                 comparer?: EqualityComparer<T, U> | true): number;
         /**
+         * Returns all elements of the collection separated by the given separator(s).
+         * 
+         * @param {U[]} {separators} One or more separator.
+         * 
+         * @return {IEnumerable<T|U>} The new sequence.
+         */
+        intersperse<U = T>(...separators: U[]): IEnumerable<T | U>;
+        /**
          * Returns the intersection between this and a second sequence.
          * 
          * @param {Sequence<T>} second The second sequence.
@@ -1773,6 +1781,31 @@ namespace Enumerable {
             }
 
             return -1;
+        }
+        /** @inheritdoc */
+        public intersperse<U = T>(...separators: U[]): IEnumerable<T | U> {
+            return from( this.intersperseInner(separators) );
+        }
+        /**
+         * @see intersperseInner()
+         */
+        protected *intersperseInner<U>(separators: U[]): Iterator<T | U> {
+            let isFirst = true;
+            for (let item of this) {
+                // separator(s)
+                if (!isFirst) {
+                    if (separators) {
+                        for (let s of separators) {
+                            yield s;
+                        }
+                    }
+                }
+                else {
+                    isFirst = false;
+                }
+
+                yield item;
+            }
         }
         /** @inheritdoc */
         public intersect(second: Sequence<T>,